Create union_example.c
this example show using of union in situation when we have some superior functionality which can  take multiple states according the settings(in this example main functionality is represent by worker, which can be official or mathematical). Union is used to store structure either of official or of mathematical. This approach allows to have only one item in worker structure, which will represent particular instance of worker and also allows use pointer on function(work function of worker), which is again set according to particular instance of worker either on work_mathematic() or on work_official()